requirements.txt
fastapi==0.115.6
uvicorn[standard]>=0.34.0
aiofiles==24.1.0
aiosqlite==0.20.0
ffmpeg-python==0.2.0
PyChromecast==14.0.4
pydantic-settings==2.7.0
pydantic==2.10.4
